Turn off ATS
App Transport Security keeps the app from running on iOS 9 with non-HTTPS servers - like the demo server.

See https://developer.apple.com/library/prerelease/ios/technotes/App-Transport-Security-Technote/ for technical information.

It's not feasible to require that every installation this app could be used with have a modern implementation of TLS, or even any implementation at all.
I have plans to make a thing to warn users if they have an insecure connection, but for right now the concern is making sure it *runs*