Change icon and launch image
Looking at this a few months later, I've come to the conclusion that
this is the better icon now.

Ungku Zoë Anysa's icon and splash screen served us great, until we
started localizing the app to other languages as part of Yosef's GSOC
project. We can't localize an icon.

Also, Apple's guidelines for icons encourage simplicity:

> Embrace simplicity. In particular, avoid cramming lots of different
images into your icon. Find a single element that captures the essence
of your app and express that element in a simple, unique shape.

https://developer.apple.com/library/ios/documentation/UserExperience/Conceptual/MobileHIG/AppIcons.html#//apple_ref/doc/uid/TP40006556-CH19-SW1